Implement the utils to convert the song into the beatmap.
